refactor(gateway): extract GatewayLifecycleMixin from run.py (slice 27 of #54962)
